+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * handle multiple classes (features) and methods (scenarios) specified in `class` argument to better align with tools which requests specific scenarios to be executed
 * support fields injection in steps classes with Hilt
+* fix [#131](https://github.com/cucumber/cucumber-android/issues/131)
+* update `cucumber-jvm` dependencies to `7.18.1`
 
 ## [7.14.0] - 2023-09-25